Refactor TextGenerator API
